Move BWE to bwe pub mod
